wifi: mac80211: don't require cipher and keylen in gtk rekey
ieee80211_add_gtk_rekey receives a keyconf as an argument, and the
cipher and keylen are taken from there to the new allocated key.
But in rekey, both the cipher and the keylen should be the same as of
the old key, so let ieee80211_add_gtk_rekey find those, so drivers won't
have to fill it in.

Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Link: https://patch.msgid.link/20250721214922.3c5c023bfae9.Ie6594ae2b4b6d5b3d536e642b349046ebfce7a5d@changeid
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
(cherry picked from commit 69fdb084355d6c0b353536024cc51aa5f7ffb62c)

[backport to 6.16: adapted iwl_mvm_gtk_rekey/iwl_mvm_d3_igtk_bigtk_rekey
 conflict resolution and converted the 6.16-only MLO rekey call site in
 iwl_mvm_mlo_gtk_rekey to the new ieee80211_gtk_rekey_add signature]